Fix #29: introduce Clean Architecture project layer separation
Split the flat src/X12Net project into three separate projects that enforce
layer boundaries at the compiler level:

- X12Net.Domain — pure types: Core, DOM, Envelopes, Schema types, Validation types
  + new X12SegmentParser (extracted from X12Reader) so DOM has no IO dependency
- X12Net.Infrastructure — IO layer: X12Reader, X12Writer, X12MemoryCapException
  X12Reader.ParseSegments now delegates to X12SegmentParser in Domain
- X12Net.Application — business logic: validators, schema validators, CLI services
  Depends on both Domain and Infrastructure

X12Net.csproj becomes the NuGet aggregator, referencing all three layers.
No circular references. All 257 tests pass unchanged.

Dependency graph: Domain ← Infrastructure ← Application ← X12Net (aggregator)
