+++ X12Net.sln
@@ -25,6 +25,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X12Net.Samples", "src\X12Ne
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X12Net.FluentValidation", "src\X12Net.FluentValidation\X12Net.FluentValidation.csproj", "{95A0728B-229A-4DA2-BF5B-0D3572B699EE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X12Net.Domain", "src\X12Net.Domain\X12Net.Domain.csproj", "{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X12Net.Infrastructure", "src\X12Net.Infrastructure\X12Net.Infrastructure.csproj", "{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X12Net.Application", "src\X12Net.Application\X12Net.Application.csproj", "{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -119,6 +125,42 @@ Global
 		{95A0728B-229A-4DA2-BF5B-0D3572B699EE}.Release|x64.Build.0 = Release|Any CPU
 		{95A0728B-229A-4DA2-BF5B-0D3572B699EE}.Release|x86.ActiveCfg = Release|Any CPU
 		{95A0728B-229A-4DA2-BF5B-0D3572B699EE}.Release|x86.Build.0 = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|x64.Build.0 = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Debug|x86.Build.0 = Debug|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|x64.ActiveCfg = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|x64.Build.0 = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|x86.ActiveCfg = Release|Any CPU
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E}.Release|x86.Build.0 = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|x64.Build.0 = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Debug|x86.Build.0 = Debug|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|x64.ActiveCfg = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|x64.Build.0 = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|x86.ActiveCfg = Release|Any CPU
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7}.Release|x86.Build.0 = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|x64.Build.0 = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Debug|x86.Build.0 = Debug|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|x64.ActiveCfg = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|x64.Build.0 = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|x86.ActiveCfg = Release|Any CPU
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -131,5 +173,8 @@ Global
 		{B21FB5AF-E97E-44C0-BFE4-721222A155AA} = {66320409-64EC-F7C5-3DEF-65E7510DAAD1}
 		{1B9D73D8-69FF-41B5-83F2-1FA04BF56C2D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{95A0728B-229A-4DA2-BF5B-0D3572B699EE} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{0864BAE3-04A9-4DCF-A556-F1CCB9B2653E} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{35DCBD04-CD63-4805-B7C0-C76DC7D06DB7} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{DBBB8CF5-5A73-454E-AF5E-35A91AB12EBF} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
