+++ src/X12Net.Application/CLI/ExceptionLogger.cs
