+++ src/X12Net.Application/CLI/IX12ToolService.cs
