+++ src/X12Net.Application/CLI/X12Tool.cs
