+++ src/X12Net.Application/CLI/X12ToolService.cs
