+++ src/X12Net.Application/Schema/X12DynamicTransaction.cs
