+++ src/X12Net.Application/Validation/Ts271Validator.cs
