+++ src/X12Net.Application/Validation/X12Validator.cs
