+++ src/X12Net.Domain/Core/X12Delimiters.cs
