+++ src/X12Net.Domain/Core/X12IsaParser.cs
