+++ src/X12Net.Domain/Core/X12Segment.cs
