+++ src/X12Net.Domain/Core/X12Token.cs
