+++ src/X12Net.Domain/Core/X12TokenType.cs
