+++ src/X12Net.Domain/Core/X12Tokenizer.cs
