+++ src/X12Net.Domain/DOM/X12FunctionalGroup.cs
