+++ src/X12Net.Domain/DOM/X12MutableSegment.cs
