+++ src/X12Net.Domain/DOM/X12Transaction.cs
