+++ src/X12Net.Domain/Envelopes/X12Envelope.cs
