+++ src/X12Net.Domain/Envelopes/X12InterchangeBuilder.cs
