+++ src/X12Net.Domain/Polyfills/IsExternalInit.cs
