+++ src/X12Net.Domain/Schema/X12SegmentSchema.cs
