+++ src/X12Net.Domain/Schema/X12TransactionSchema.cs
