+++ src/X12Net.Domain/Validation/X12ErrorCode.cs
