+++ src/X12Net.Domain/Validation/X12ValidationError.cs
