+++ src/X12Net.Domain/Validation/X12ValidationResult.cs
