+++ src/X12Net.Domain/Validation/X12ValidationRule.cs
