+++ src/X12Net.Infrastructure/IO/X12MemoryCapException.cs
