+++ src/X12Net.Infrastructure/IO/X12Writer.cs
